requirements.txt
Flask==2.0.1
selenium==4.0.0
camelot-py==0.10.1
pandas==1.3.0